Main/Lambda/batch_processor_requirements.txt
python-dotenv==1.0.0
pandas==1.5.3
xgboost==2.0.0
kafka-python==2.0.2
hdfs==2.7.0
pymongo==4.5.0 